Migrating to ivec3s everywhere

+++ buildings/building_plan.py
@@ -1,14 +1,14 @@
 from dataclasses import dataclass
 from structures.grid import Grid
-from structures.types import vec3
 from buildings.cell import Cell
+from gdpc.vector_tools import ivec3
 
 @dataclass
 class BuildingPlan:
     grid : Grid
-    shape : list[vec3]
-    cells : dict[vec3, Cell]
+    shape : list[ivec3]
+    cells : dict[ivec3, Cell]
 
-    def __init__(self, shape : list[vec3]) -> None:
+    def __init__(self, shape : list[ivec3]) -> None:
         self.shape = shape
         self.cells = {pt : Cell(position=pt, plan=self) for pt in shape}
