+++ buildings/walls/wall.py
@@ -2,7 +2,7 @@ from data.asset import Asset
 from data.asset import Asset
 from gdpc.editor import Editor
 from structures.grid import Grid
-from structures.types import vec3
+from gdpc.vector_tools import ivec3
 
 # Abstract class for walls
 class Wall(Asset):
@@ -10,7 +10,7 @@ class Wall(Asset):
         self, 
         editor: Editor, 
         grid: Grid, 
-        grid_coordinate: vec3, 
+        grid_coordinate: ivec3, 
         facing: str = None):
             pass
